+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -76,8 +76,12 @@ public enum NoteTaskIdentifier
     [Description("commercial_transfer_agreement")]
     [DisplayDescription("Commercial transfer agreement")]
     CommercialTransferAgreement = 18,
+
+    [Description("main_contact")]
+    [DisplayDescription("Confirm the main contact")]
+    MainContact = 19,
     
     [Description("master_funding_agreement")]
     [DisplayDescription("Master funding agreement")]
-    MasterFundingAgreement = 19
+    MasterFundingAgreement = 20
 }