+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ProjectsController/TaskDataControllerTests.cs
@@ -11,7 +11,6 @@ using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
 using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
-using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
 
 namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
 {
@@ -716,7 +715,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
 
             await dbContext.SaveChangesAsync();
             command.TaskDataId = new TaskDataId { Value = taskData.Id.Value }; 
-            command.AcademyOpenedDate = new DateTime(2025, 1, 1);
+            command.AcademyOpenedDate = new DateTime(2025, 1, 1, 0, 0, 0, DateTimeKind.Utc);
 
             // Act
             await tasksDataClient.UpdateConfirmAcademyOpenedDateTaskAsync(command, default);
@@ -816,6 +815,109 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.True(existingTaskData.ChurchSupplementalAgreementSignedSecretaryState);
 
         }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateMainContactTaskAsync_ShouldUpdate_TransferTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateMainContactTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>(); 
+            var testUser = await dbContext.Users.FirstOrDefaultAsync();
+            var establishment = fixture.Create<Domain.Entities.GiasEstablishment>();
+            var localAuthority = fixture.Create<Domain.Entities.LocalAuthority>();
+            await dbContext.LocalAuthorities.AddAsync(localAuthority);
+            await dbContext.GiasEstablishments.AddAsync(establishment);
+            await dbContext.LocalAuthorities.AddAsync(localAuthority);
+            await dbContext.GiasEstablishments.AddAsync(establishment);
+            var project = fixture.Customize(
+                new ProjectCustomization()
+                {
+                    Id = new Domain.ValueObjects.ProjectId(Guid.NewGuid()),
+                    Urn = establishment.Urn ?? new Domain.ValueObjects.Urn(123456),
+                    LocalAuthorityId = localAuthority.Id,
+                    RegionalDeliveryOfficerId = testUser!.Id,
+                    Type = Domain.Enums.ProjectType.Transfer,
+                    AllConditionsMet = false,
+                    TasksDataType = Domain.Enums.TaskType.Transfer,
+                }
+                ).Create<Domain.Entities.Project>();
+            dbContext.Projects.Add(project);
+
+            await dbContext.SaveChangesAsync();
+            command.ProjectId = new ProjectId { Value = project.Id.Value };
+            command.MainContactId = new ContactId { Value = Guid.NewGuid() };
+
+            // Act
+            await tasksDataClient.UpdateMainContactTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingProject = await dbContext.Projects.SingleOrDefaultAsync(x => x.Id == project.Id);
+            Assert.NotNull(existingProject);
+            Assert.NotNull(existingProject.MainContactId);
+        }
+
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateExternalStakeholderKickOffTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateExternalStakeholderKickOffTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+            var testUser = await dbContext.Users.FirstOrDefaultAsync();
+            var establishment = fixture.Create<Domain.Entities.GiasEstablishment>();
+            var localAuthority = fixture.Create<Domain.Entities.LocalAuthority>();
+            await dbContext.LocalAuthorities.AddAsync(localAuthority);
+            await dbContext.GiasEstablishments.AddAsync(establishment);
+            await dbContext.LocalAuthorities.AddAsync(localAuthority);
+            await dbContext.GiasEstablishments.AddAsync(establishment);
+            var project = fixture.Customize(
+                new ProjectCustomization()
+                {
+                    Id = new Domain.ValueObjects.ProjectId(Guid.NewGuid()),
+                    Urn = establishment.Urn ?? new Domain.ValueObjects.Urn(123456),
+                    LocalAuthorityId = localAuthority.Id,
+                    RegionalDeliveryOfficerId = testUser!.Id,
+                    Type = Domain.Enums.ProjectType.Conversion,
+                    TasksDataId = taskData.Id,
+                    AllConditionsMet = false,
+                    TasksDataType = Domain.Enums.TaskType.Conversion,
+                }
+                ).Create<Domain.Entities.Project>();
+            dbContext.Projects.Add(project); 
+
+            await dbContext.SaveChangesAsync();
+            command.ProjectId = new ProjectId { Value = project.Id.Value };
+            command.StakeholderKickOffSetupMeeting = false;
+            command.StakeholderKickOffIntroductoryEmails = true; 
+            command.SignificantDate = new DateTime(2025, 1, 1,0,0,0, DateTimeKind.Utc);
+            command.UserEmail = testUser.Email;
+            // Act
+            await tasksDataClient.UpdateExternalStakeholderKickOffTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.StakeholderKickOffIntroductoryEmails);
+            Assert.False(existingTaskData.StakeholderKickOffSetupMeeting);
+
+            var existingProject = await dbContext.Projects.SingleOrDefaultAsync(x => x.Id == project.Id);
+            Assert.NotNull(existingProject);
+            Assert.NotNull(existingProject.SignificantDate);
+        }
 
         [Theory]
         [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
@@ -893,6 +995,42 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.True(existingTaskData.CommercialTransferAgreementQuestionsChecked);
             Assert.False(existingTaskData.CommercialTransferAgreementSaved);
             Assert.False(existingTaskData.CommercialTransferAgreementSigned);
+        }
+
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateLandQuestionnaireTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateLandQuestionnaireTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.Received = true;
+            command.Signed = true;
+            command.Saved = false;
+            command.Cleared = false;
+
+            // Act
+            await tasksDataClient.UpdateLandQuestionnaireTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.False(existingTaskData.LandQuestionnaireCleared);
+            Assert.True(existingTaskData.LandQuestionnaireReceived);
+            Assert.True(existingTaskData.LandQuestionnaireSigned);
+            Assert.False(existingTaskData.LandQuestionnaireSaved);
         }